requirements.txt
playwright==1.49.1
requests==2.32.3
pyotp==2.9.0
python-dotenv==1.0.1
flask>=3.0
flask-sqlalchemy>=3.1
flask-login>=0.6
flask-sock>=0.7
flask-migrate>=4.0
websockets>=12.0
psycopg2-binary>=2.9
celery[redis]>=5.4
redis>=5.0
anthropic>=0.40
pdfplumber>=0.11
beautifulsoup4>=4.12
paramiko>=3.0
